GCS 3.0 Prototype Nearly Finalized
The only function critical change needed is the implementation of
calibrations which will require referencing the older LabView code.

Other changes needed for future development include: implementation of
graphing sensor data, showing sensor readings on the P&ID instead of in
a table, and minor aesthetic changes.

Other minor changes should be expected based on testing and consultation
with team members.

Documentation is incomplete, this will need to be finalized.

To open the executable, go to GCS_3.0\bin\Debug\GCS_3.0.exe. In the
future, once the application is finalized, an installer can be made.