Functional Finalization of GCS 3.0 Prototype
This commit includes the implementation of the calibration values for
the different sensors and the use of the Lunar Lion icon for the .exe
file. With that, all the functional aspects of the prototype GCS 3.0
have been completed (this, of course, excludes graphs and a P&ID with
sensor readings instead of a table).