Interface for Designing Flight Profiles
A new "Flight Profile Designer" folder was made in which the files
relating to a new application for designing flight profiles reside.
Within this folder there is one folder for the raw, Visual Studio files
and another zip folder for the installation.

To install the application, the zip folder must be extracted and the
setup.exe must be run. Windows SmartScreen may throw a warning because I
do not hold an official developer license but SmartScreen will allow the
user to bypass the warning.