Formatting of Datalog Corrected
The formatting of the datalog was previously incorrect in two ways.

Firstly, the column headings for the servo motor values did not use the
motor names but instead used motor values. Line 480 in
MainWindow.xaml.cs was changed to use the "name" property of servo motor
instead of the "pwm" property.

Secondly, there was a redundant logging of servo motor values because of
an unnecessary for-each loop which logged servo motor values a second
time for each iteration and this loop was removed.